# Improve the performance of bayesian_blocks (binned or unbinned events)